Moved Token to graql.util.Token, and fixed all build files
