+++ server/src/graql/query/GraqlGet.java
@@ -21,6 +21,7 @@ package grakn.core.graql.query;
 import grakn.core.graql.query.builder.AggregateBuilder;
 import grakn.core.graql.query.statement.Variable;
 import graql.exception.GraqlException;
+import graql.util.Token;
 
 import javax.annotation.CheckReturnValue;
 import java.util.LinkedHashSet;
