+++ server/src/graql/query/pattern/Negation.java
@@ -19,7 +19,7 @@
 package grakn.core.graql.query.pattern;
 
 import grakn.core.graql.query.Graql;
-import grakn.core.graql.query.Token;
+import graql.util.Token;
 import grakn.core.graql.query.statement.Statement;
 import grakn.core.graql.query.statement.Variable;
