+++ server/src/graql/query/property/DataTypeProperty.java
@@ -18,7 +18,7 @@
 
 package grakn.core.graql.query.property;
 
-import grakn.core.graql.query.Token;
+import graql.util.Token;
 import grakn.core.graql.query.statement.StatementType;
 
 /**
