+++ server/test/graql/query/predicate/RegexPredicateTest.java
@@ -19,7 +19,7 @@
 package grakn.core.graql.query.predicate;
 
 import grakn.core.graql.internal.executor.property.ValueExecutor;
-import grakn.core.graql.query.Token;
+import graql.util.Token;
 import grakn.core.graql.query.property.ValueProperty;
 import org.junit.Test;
