+++ test-integration/graql/query/BUILD
@@ -37,8 +37,9 @@ java_test(
     test_class = "grakn.core.graql.query.GraqlDefineIT",
     srcs = ["GraqlDefineIT.java"],
     deps = [
-        "//common",
-        "//server",
+        "//common:common",
+        "//graql/java:graql",
+        "//server:server",
         "//test-integration/graql/graph:movie-graph",
         "//test-integration/rule:grakn-test-server",
         "//test-integration/util:graql-test-util",
@@ -87,7 +88,8 @@ java_test(
     test_class = "grakn.core.graql.query.GraqlUndefineIT",
     srcs = ["GraqlUndefineIT.java"],
     deps = [
-        "//server",
+        "//graql/java:graql",
+        "//server:server",
         "//test-integration/graql/graph:movie-graph",
         "//test-integration/rule:grakn-test-server",
         "//test-integration/util:graql-test-util",
