+++ test-integration/graql/query/GraqlUndefineIT.java
@@ -35,6 +35,7 @@ import grakn.core.server.Transaction;
 import grakn.core.server.exception.TransactionException;
 import grakn.core.server.session.SessionImpl;
 import grakn.core.server.session.TransactionOLTP;
+import graql.util.Token;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.ClassRule;
