+++ test-integration/server/AttributeDeduplicatorIT.java
@@ -23,7 +23,7 @@ import grakn.core.graql.concept.Label;
 import grakn.core.graql.internal.Schema;
 import grakn.core.graql.query.Graql;
 import grakn.core.graql.query.GraqlInsert;
-import grakn.core.graql.query.Token;
+import graql.util.Token;
 import grakn.core.rule.GraknTestServer;
 import grakn.core.server.deduplicator.AttributeDeduplicator;
 import grakn.core.server.deduplicator.KeyspaceIndexPair;
